# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Exam API docs",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "paths": {
    "/Articles": {
      "get": {
        "summary": "Get all articles",
        "responses": {
          "200": {
            "description": "A list of articles",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Article"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/Articles/{id}": {
      "get": {
        "summary": "Retrieve an article",
        "description": "Get a specific article by its ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the article to retrieve",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Article retrieved successfully",
            "schema": {
              "$ref": "#/components/schemas/Article"
            }
          },
          "404": {
            "description": "Article not found"
          },
          "500": {
            "description": "An error occurred while getting the article"
          }
        }
      }
    },
    "/{id}/comments": {
      "post": {
        "summary": "Add a new comment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "requestBody": {
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "user": {
                        "type": "string"
                      },
                      "text": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "user",
                      "text"
                    ]
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Comment added successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Comment"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "get": {
        "summary": "Get comments for an article",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of comments",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Comment"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Article not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/register": {
      "post": {
        "summary": "Register a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The created user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      }
    },
    "/login": {
      "post": {
        "summary": "Log in a user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Login"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The logged in user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Article": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "The title of the article"
          },
          "link": {
            "type": "string",
            "description": "The link to the article"
          },
          "summary": {
            "type": "string",
            "description": "The summary of the article"
          },
          "image": {
            "type": "string",
            "description": "The image URL of the article"
          }
        }
      },
      "Comment": {
        "type": "object",
        "properties": {
          "user": {
            "type": "string",
            "format": "uuid",
            "description": "The ID of the user who posted the comment"
          },
          "text": {
            "type": "string",
            "description": "The text of the comment"
          },
          "article": {
            "type": "string",
            "format": "uuid",
            "description": "The ID of the article the comment is associated with"
          }
        },
        "required": [
          "user",
          "text",
          "article"
        ]
      },
      "User": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "The unique username of the user"
          },
          "password": {
            "type": "string",
            "description": "The password of the user"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "The unique email of the user"
          }
        },
        "required": [
          "username",
          "password",
          "email"
        ]
      },
      "Login": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "The username of the user"
          },
          "password": {
            "type": "string",
            "description": "The password of the user"
          }
        },
        "required": [
          "username",
          "password"
        ]
      }
    }
  }
}